refactor(cognito): split service/auth.rs by flow (2720 -> 402 mod + 7 submodules)
Audit 2026-05-19 P9 continuation. Cognito auth handlers split:

- initiate.rs    781  (initiate_auth / *_user_password_auth / *_custom_auth /
                     *_refresh_token_auth)
- challenges.rs  602  (respond_to_auth_challenge / handle_auth_challenge_response /
                     respond_new_password_required / respond_custom_challenge /
                     custom_auth_issue_tokens / custom_challenge_issue_tokens)
- admin.rs       372  (admin_initiate_auth / admin_auth_lookup / admin_auth_verify /
                     admin_respond_to_auth_challenge / admin_confirm_sign_up /
                     admin_reset_user_password / admin_user_global_sign_out)
- password.rs    285  (change_password / forgot_password / confirm_forgot_password)
- signup.rs      261  (sign_up / confirm_sign_up)
- compromised.rs  52  (evaluate_compromised_credentials)
- signout.rs      39  (global_sign_out)
- mod.rs         402  (AuthFlowParams parsing + tests)

320 unit tests pass.
